update: get compiling with current libs, switch to Text.
This huge set of changes is primarily composed of two things:
1. Heist switched to using MapSyntax via do blocks, instead of lists of
tuples.
2. Some libraries switched away from ByteString to Text, and instead of
trying to figure out which had and which should use which, I essentially
switched everything to use Text instead.
3. Finally, I set this up using Rivet (note that email-postmark has to
be updated as well) and used cabal's freeze to write out the current
working dependencies (as a result, the cabal file has no bounds specified).